pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai_gateway"
version = "0.1.0"
description = "Code generation provided by GitLab.com"
authors = ["GitLab AI Assisted"]
readme = "README.md"
packages = [{ include = "ai_gateway" }]

[tool.poetry.dependencies]
python = "~3.9"
fastapi = "^0.82.0"
dependency-injector = "^4.41.0"
uvicorn = { extras = ["standard"], version = "^0.20.0" }
requests = "^2.28.1"
numpy = "^1.24.1"
python-dotenv = "^0.21.0"
regex = "^2022.10.31"
detect-secrets = "^1.4.0"
fastapi-health = "^0.4.0"
google-cloud-aiplatform = "^1.25.0"
structlog = "^23.1.0"
asgi-correlation-id = "^4.1.0"
prometheus-fastapi-instrumentator = "^6.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
starlette-context = "^0.3.6"
google-cloud-profiler = "^4.0.0"
py-grpc-prometheus = "^0.7.0"
tree-sitter = "^0.20.1"
transformers = "^4.30.0"
sentencepiece = "^0.1.99"
snowplow-tracker = "^1.0.1"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
responses = "^0.23.1"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"

[tool.pytest.ini_options]
addopts = "--disable-warnings --cov=ai_gateway --cov-report=term --cov-report=html"
testpaths = ["tests/codesuggestions"]
filterwarnings = ["error::UserWarning"]

[tool.poetry.scripts]
ai_gateway = "ai_gateway.app:main"

[tool.isort]
profile = "black"
skip = "scripts/vendor"
